avfilter/hwupload: wait for all pending transfers on uninit()
Uploads can be asynchronous, so we need to wait for them on unit to prevent
a cyclic reference from keeping the hwfc alive indefinitely, since there is
no other way for the hwfc to know when prior commands should be cleaned up.

In theory, it might be possible for this to be signalled backwards by
downstream consumers of the corresponding AVFrames, but this is very nontrivial
as there may be multiple, or even no downstream consumers, such as in the
degenerate case of:

 ffmpeg ... -vf hwupload -f null -

Waiting for outstanding transfers to have completed explicitly in uninit
is the only solution that is both robust, simple, and comes without a
performance penalty.