avfilter/vf_libplacebo: always pass through SAR when stretching
The current logic assumes `s->reset_sar` always resets the SAR, but this
is only true if `force_oar` is actually true as well, meaning that in the
case of `s->reset_sar && !force_oar`, the SAR is incorrectly set to 1:1.

While we could fix this by also forcing `force_oar` for multiple inputs,
that would have unintended consequences when the user explicitly sets the
desired resolution. Instead, pass-through the SAR in this case; matching
the intended behavior of inheriting the SAR from the base layer input.